add distinct sudo/non-sudo and seed_derivation modes fix #98
