+++ app/templates/index.html
@@ -16,6 +16,7 @@
     </p>
 </body>
 <footer>
-    Version: {{ version }}
+    <p>Version: {{ version }}</p>
+    <p>Sudo Mode: {{ sudo_mode }}</p>
 </footer>
 </html>
